Split out when steps by handler
