+++ systemTests/testScenarios/when/delete_file_when_steps.py
@@ -1,0 +1,18 @@
+from testScenarios.when.base_when_step import BaseWhenStep
+
+
+class DeleteFileWhenSteps(BaseWhenStep):
+    def authorised_delete_file_is_called_for_simple_file_in_s3(self):
+        handler = 'deletefile'
+        body = {'bucket': self._context.standard_bucket_name, 'file': 'testA.txt'}
+        self._make_authorised_request(handler, body)
+
+    def unauthorised_delete_file_called_for_whole_directory(self):
+        handler = 'deletefile'
+        body = {'bucket': self._context.standard_bucket_name, 'file': 'testA.txt'}
+        self._make_unauthorised_request(handler, body)
+
+    def delete_file_called_with_no_authorisation_credentials(self):
+        handler = 'deletefile'
+        body = {'bucket': self._context.standard_bucket_name, 'file': 'testA.txt'}
+        self._make_request_with_no_authorisation_credentials(handler, body)
